+++ packages/antdv1-enhance/packages/action/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @yuyi919/antdv1-plus-action
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @yuyi919/antdv1-plus-theme@0.0.4
+  - @yuyi919/vue-jsx-factory@0.0.10
+  - @yuyi919/antdv1-plus-shared@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes